Make MaxSAT algorithm use new anti-unification code.
