+++ core/src/main/scala/at/logic/gapt/grammars/deltatable/deltatable2.scala
@@ -2,85 +2,13 @@ package at.logic.gapt.grammars.deltatable
 
 import at.logic.gapt.cutintro.{ GrammarFindingMethod, DeltaTableMethod }
 import at.logic.gapt.expr._
-import at.logic.gapt.grammars.VectTratGrammar
+import at.logic.gapt.grammars.{ antiUnifier, antiUnifier1, VectTratGrammar }
 import at.logic.gapt.utils.time
 
 import scala.collection.mutable
 
 object deltaTable {
   type Row = Set[( LambdaExpression, Set[LambdaExpression] )]
-  type SimpleGrammar = ( LambdaExpression, Set[Substitution] )
-
-  def antiUnifier( a: LambdaExpression, b: LambdaExpression ): ( LambdaExpression, collection.Map[Var, LambdaExpression], collection.Map[Var, LambdaExpression] ) = {
-    val vars = mutable.Map[( LambdaExpression, LambdaExpression ), Var]()
-    val subst1 = mutable.Map[Var, LambdaExpression]()
-    val subst2 = mutable.Map[Var, LambdaExpression]()
-
-    var i = 0
-    def au( a: LambdaExpression, b: LambdaExpression ): LambdaExpression = {
-      val Apps( fa, as ) = a
-      val Apps( fb, bs ) = b
-      if ( fa == fb ) {
-        fa( ( as, bs ).zipped map au: _* )
-      } else {
-        vars.getOrElseUpdate( a -> b, {
-          i = i + 1
-          val v = Var( s"x$i", a.exptype )
-          subst1( v ) = a
-          subst2( v ) = b
-          v
-        } )
-      }
-    }
-
-    ( au( a, b ), subst1, subst2 )
-  }
-
-  def antiUnifier1( a: LambdaExpression, b: LambdaExpression ): ( LambdaExpression, collection.Map[Var, LambdaExpression], collection.Map[Var, LambdaExpression] ) = {
-    def au1FromAU( au: LambdaExpression ): ( LambdaExpression, Option[LambdaExpression] ) = au match {
-      case Var( _, ty ) => ( Var( "x", ty ), Some( au ) )
-      case c: Const     => ( c, None )
-      case Apps( f, as ) =>
-        val ( as_, s_ ) = as.map( au1FromAU ).unzip
-        if ( s_.flatten.distinct.size <= 1 ) {
-          f( as_ : _* ) -> s_.flatten.headOption
-        } else {
-          Var( "x", au.exptype ) -> Some( au )
-        }
-    }
-
-    val ( au, subst1, subst2 ) = antiUnifier( a, b )
-    au1FromAU( au ) match {
-      case ( au1, None ) => ( au1, Map(), Map() )
-      case ( au1, Some( repl ) ) =>
-        val v = freeVariables( au1 ).head
-        ( au1, Map( v -> Substitution( subst1 )( repl ) ), Map( v -> Substitution( subst2 )( repl ) ) )
-    }
-  }
-
-  def antiUnifier1b( a: LambdaExpression, b: LambdaExpression ): ( LambdaExpression, collection.Map[Var, LambdaExpression], collection.Map[Var, LambdaExpression] ) = {
-    def au( a: LambdaExpression, b: LambdaExpression ): ( LambdaExpression, Option[( LambdaExpression, LambdaExpression )] ) = {
-      val Apps( fa, as ) = a
-      val Apps( fb, bs ) = b
-      if ( fa == fb ) {
-        val ( as_, s_ ) = ( as, bs ).zipped.map( au ).unzip
-        if ( s_.flatten.distinct.size <= 1 ) {
-          ( fa( as_ : _* ), s_.flatten.headOption )
-        } else {
-          ( Var( "x", a.exptype ), Some( ( a, b ) ) )
-        }
-      } else {
-        ( Var( "x", a.exptype ), Some( ( a, b ) ) )
-      }
-    }
-
-    au( a, b ) match {
-      case ( au1, None ) => ( au1, Map(), Map() )
-      case ( au1, Some( ( substA, substB ) ) ) =>
-        val v = freeVariables( au1 ).head
-        ( au1, Map( v -> substA ), Map( v -> substB ) )
-    }
-  }
 
   def createTable( termSet: Set[LambdaExpression], maxArity: Option[Int] = None, singleVariable: Boolean = false ): Map[Set[Substitution], Row] = {
     // invariant:  deltatable(S) contains (u,T)  ==>  u S = T  &&  |S| = |T|
